+++ ACA.DeliverySystem.Data/Repository/UserRepository.cs
@@ -17,7 +17,7 @@ namespace ACA.DeliverySystem.Data.Repository
 
         public async Task<User> GetById(int id, CancellationToken token)
         {
-            return await _context.Users.FirstOrDefaultAsync(x => x.Id == id);
+            return await _context.Users.SingleOrDefaultAsync(x => x.Id == id);
         }
 
         public async Task Add(User user, CancellationToken token)
@@ -28,13 +28,17 @@ namespace ACA.DeliverySystem.Data.Repository
         public async Task Update(User user, CancellationToken token)
         {
             _context.Users.Update(user);
-
         }
 
-        public async Task Delete(int id, CancellationToken token)
+        public async Task<OperationResult> Delete(int id, CancellationToken token)
         {
-            var user = await _context.Users.SingleAsync(e => e.Id == id, token);
+            var user = await _context.Users.SingleOrDefaultAsync(e => e.Id == id, token);
+            if (user == null)
+            {
+                return OperationResult.Error($"User with id {id} not found.", ErrorType.NotFound);
+            }
             _context.Users.Remove(user);
+            return OperationResult.Ok();
         }
 
         public async Task<IEnumerable<Order>> GetUserOrders(int id, CancellationToken token)
